Fix the size_t to int warning

+++ taskflow/graph/basic_taskflow.hpp
@@ -463,7 +463,7 @@ void BasicTaskflow<E>::Closure::operator () () const {
         node->_dependents.pop_back();
       }
     }
-    node->_num_dependents = node->_dependents.size();
+    node->_num_dependents = static_cast<int>(node->_dependents.size());
     node->unset_spawned();
   }
